+++ src/snowflake/snowpark/_internal/compiler/large_query_breakdown.py
@@ -51,7 +51,9 @@ from snowflake.snowpark._internal.compiler.telemetry_constants import (
 )
 from snowflake.snowpark._internal.compiler.utils import (
     TreeNode,
+    extract_child_from_with_query_block,
     is_active_transaction,
+    is_with_query_block,
     replace_child,
     update_resolvable_node,
 )
@@ -271,7 +273,7 @@ class LargeQueryBreakdown:
                 )
                 break
 
-            partition = self._get_partitioned_plan(root, child)
+            partition = self._get_partitioned_plan(child)
             plans.append(partition)
             complexity_score = get_complexity_score(root)
 
@@ -356,7 +358,7 @@ class LargeQueryBreakdown:
             current_node_validity_statistics,
         )
 
-    def _get_partitioned_plan(self, root: TreeNode, child: TreeNode) -> SnowflakePlan:
+    def _get_partitioned_plan(self, child: TreeNode) -> SnowflakePlan:
         """This method takes cuts the child out from the root, creates a temp table plan for the
         partitioned child and returns the plan. The steps involved are:
 
@@ -375,7 +377,9 @@ class LargeQueryBreakdown:
                 [temp_table_name],
                 None,
                 SaveMode.ERROR_IF_EXISTS,
-                child,
+                extract_child_from_with_query_block(child)
+                if is_with_query_block(child)
+                else child,
                 table_type="temp",
                 creation_source=TableCreationSource.LARGE_QUERY_BREAKDOWN,
             )